+++ src/components/Button/index.tsx
@@ -1,6 +1,6 @@
 import {ButtonHTMLAttributes} from 'react';
 
-import '../styles/button.scss';
+import './styles.scss';
 
 type ButtonProps  = ButtonHTMLAttributes<HTMLButtonElement> & {
   isOutlined?: boolean
