+++ src/pages/AdminRoom/index.tsx
@@ -1,18 +1,18 @@
 import { useHistory, useParams } from 'react-router-dom';
 
-import deleteImg from '../assets/images/delete.svg';
-import logoImg from '../assets/images/logo.svg';
-import checkImg from '../assets/images/check.svg';
-import answerImg from '../assets/images/answer.svg';
+import deleteImg from '../../assets/images/delete.svg';
+import logoImg from '../../assets/images/logo.svg';
+import checkImg from '../../assets/images/check.svg';
+import answerImg from '../../assets/images/answer.svg';
 
-import { Button } from '../components/Button';
-import { Question } from '../components/Question';
-import { RoomCode } from '../components/RoomCode';
-// import { useAuth } from '../hooks/useAuth';
-import { useRoom } from '../hooks/useRoom';
-import { database } from '../services/firebase';
+import { Button } from '../../components/Button';
+import { Question } from '../../components/Question';
+import { RoomCode } from '../../components/RoomCode';
+// import { useAuth } from '../../hooks/useAuth';
+import { useRoom } from '../../hooks/useRoom';
+import { database } from '../../services/firebase';
 
-import '../styles/room.scss';
+import './styles.scss';
 
 type RoomParams = {
   id: string;
