+++ src/pages/Home/index.tsx
@@ -1,16 +1,15 @@
+import { useState, FormEvent } from 'react';
 import { useHistory } from 'react-router-dom';
-import { FormEvent } from 'react';
 
-import illustrationImg from '../assets/images/illustration.svg';
-import logoImg from '../assets/images/logo.svg';
-import googleIconImg from '../assets/images/google-icon.svg';
+import illustrationImg from '../../assets/images/illustration.svg';
+import logoImg from '../../assets/images/logo.svg';
+import googleIconImg from '../../assets/images/google-icon.svg';
 
-import '../styles/auth.scss';
+import './styles.scss';
 
-import { Button } from '../components/Button';
-import { useAuth } from '../hooks/useAuth';
-import { useState } from 'react';
-import { database } from '../services/firebase';
+import { Button } from '../../components/Button';
+import { useAuth } from '../../hooks/useAuth';
+import { database } from '../../services/firebase';
 
 export function Home() {
   const history = useHistory();
